Refactor: migrate deployments to generated cached store
